requirements.txt
fastapi
pydantic
python-dotenv
python-multipart
requests
starlette
uvicorn
sqlalchemy